json: err tests
