+++ facet-json/tests/integration/read/rename.rs
@@ -1,0 +1,463 @@
+/// Basic deserialization with renamed fields
+#[test]
+fn test_field_rename_deserialization() {
+    facet_testhelpers::setup();
+
+    #[derive(Facet, Debug, PartialEq)]
+    struct Greetings {
+        #[facet(rename = "bonjour")]
+        hello: String,
+
+        #[facet(rename = "au_revoir")]
+        goodbye: String,
+    }
+
+    let json = r#"{"bonjour":"monde","au_revoir":"world"}"#;
+
+    let result: Greetings = match from_str(json) {
+        Ok(s) => s,
+        Err(e) => panic!("Error deserializing JSON: {}", e),
+    };
+
+    assert_eq!(result.hello, "monde");
+    assert_eq!(result.goodbye, "world");
+}
+
+/// Round-trip serialization then deserialization with a renamed field
+#[cfg(feature = "std")]
+#[test]
+fn test_field_rename_roundtrip() {
+    facet_testhelpers::setup();
+
+    #[derive(Facet, Debug, PartialEq)]
+    struct Greetings {
+        #[facet(rename = "bonjour")]
+        hello: String,
+    }
+
+    let original = Greetings {
+        hello: "monde".to_string(),
+    };
+
+    let json = facet_json::to_string(&original);
+    assert_eq!(json, r#"{"bonjour":"monde"}"#);
+
+    let roundtrip: Greetings = from_str(&json).unwrap();
+    assert_eq!(original, roundtrip);
+}
+
+/// Deserialization with common naming conventions (kebab-case, snake_case, camelCase)
+#[test]
+fn test_field_rename_common_case_styles() {
+    facet_testhelpers::setup();
+
+    #[derive(Facet, Debug, PartialEq)]
+    struct SpecialNames {
+        #[facet(rename = "kebab-case")]
+        kebab_case: String,
+
+        #[facet(rename = "snake_case")]
+        original_snake: String,
+
+        #[facet(rename = "camelCase")]
+        camel_case: String,
+    }
+
+    let json = r#"{"kebab-case":"dash","snake_case":"underscore","camelCase":"hump"}"#;
+
+    let result: SpecialNames = match from_str(json) {
+        Ok(s) => s,
+        Err(e) => panic!("Error deserializing JSON: {}", e),
+    };
+
+    assert_eq!(result.kebab_case, "dash");
+    assert_eq!(result.original_snake, "underscore");
+    assert_eq!(result.camel_case, "hump");
+}
+
+/// Serialization and deserialization with special symbol characters in field name
+#[test]
+#[cfg(feature = "std")]
+fn test_field_rename_with_symbol_chars_name() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct SpecialCharsName {
+        #[facet(rename = "@#$%^&")]
+        special_chars: String,
+    }
+
+    let test_struct = SpecialCharsName {
+        special_chars: "special value".to_string(),
+    };
+
+    let json = facet_json::to_string(&test_struct);
+    assert_eq!(json, r#"{"@#$%^&":"special value"}"#);
+
+    let roundtrip: SpecialCharsName = facet_json::from_str(&json).unwrap();
+    assert_eq!(test_struct, roundtrip);
+}
+
+/// Serialization and deserialization with Unicode characters in field name (emoji)
+#[test]
+#[cfg(feature = "std")]
+fn test_field_rename_with_unicode_name_emoji() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct EmojiCharsName {
+        #[facet(rename = "🏀")]
+        ball: String,
+    }
+
+    let test_struct = EmojiCharsName {
+        ball: "🏆".to_string(),
+    };
+
+    let json = facet_json::to_string(&test_struct);
+    assert_eq!(json, r#"{"🏀":"🏆"}"#);
+
+    let roundtrip: EmojiCharsName = facet_json::from_str(&json).unwrap();
+    assert_eq!(test_struct, roundtrip);
+}
+
+/// Serialization and deserialization with Unicode characters in field name (Euro sign)
+#[test]
+#[cfg(feature = "std")]
+fn test_field_rename_with_unicode_name_special_signs() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct EmojiCharsName {
+        #[facet(rename = "€℮↑→↓↔↕")]
+        special_chars: String,
+    }
+
+    let test_struct = EmojiCharsName {
+        special_chars: "...".to_string(),
+    };
+
+    let json = facet_json::to_string(&test_struct);
+    assert_eq!(json, r#"{"€℮↑→↓↔↕":"..."}"#);
+
+    let roundtrip: EmojiCharsName = facet_json::from_str(&json).unwrap();
+    assert_eq!(test_struct, roundtrip);
+}
+
+/// Serialization and deserialization with numeric field name
+#[cfg(feature = "std")]
+#[test]
+fn test_field_rename_with_numeric_name() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct NumericName {
+        #[facet(rename = "123")]
+        numeric_name: i32,
+    }
+
+    let test_struct = NumericName { numeric_name: 42 };
+
+    let json = facet_json::to_string(&test_struct);
+    assert_eq!(json, r#"{"123":42}"#);
+
+    let roundtrip: NumericName = facet_json::from_str(&json).unwrap();
+    assert_eq!(test_struct, roundtrip);
+}
+
+/// Serialization and deserialization with empty field name
+#[cfg(feature = "std")]
+#[test]
+#[ignore]
+fn test_field_rename_with_empty_name() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct EmptyName {
+        #[facet(rename = "")]
+        empty_name: bool,
+    }
+
+    let test_struct = EmptyName { empty_name: true };
+
+    let json = facet_json::to_string(&test_struct);
+    assert_eq!(json, r#"{"":true}"#);
+
+    let roundtrip: EmptyName = facet_json::from_str(&json).unwrap();
+    assert_eq!(test_struct, roundtrip);
+}
+
+/// Serialization and deserialization of renamed enum variants (unit and tuple variants)
+#[cfg(feature = "std")]
+#[test]
+#[ignore]
+fn test_enum_variant_rename() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    #[repr(u8)]
+    enum Color {
+        #[facet(rename = "lime")]
+        Green,
+
+        #[facet(rename = "cyan")]
+        Blue(u8),
+    }
+
+    // Test unit variant with rename
+    let green = Color::Green;
+    let json = facet_json::to_string(&green);
+    assert_eq!(json, r#""lime""#);
+    let roundtrip: Color = facet_json::from_str(&json).unwrap();
+    assert_eq!(green, roundtrip);
+
+    // Test tuple variant with rename
+    let blue = Color::Blue(255);
+    let json = facet_json::to_string(&blue);
+    assert_eq!(json, r#"{"cyan":255}"#);
+    let roundtrip: Color = facet_json::from_str(&json).unwrap();
+    assert_eq!(blue, roundtrip);
+}
+
+/// Serialization and deserialization of renamed fields in struct enum variants
+#[cfg(feature = "std")]
+#[test]
+#[ignore]
+fn test_enum_struct_variant_field_rename() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    #[repr(u8)]
+    enum Message {
+        #[facet(rename = "success")]
+        Success {
+            #[facet(rename = "message")]
+            msg: String,
+
+            #[facet(rename = "code")]
+            status_code: u16,
+        },
+
+        #[facet(rename = "error")]
+        Error {
+            #[facet(rename = "errorMessage")]
+            msg: String,
+
+            #[facet(rename = "errorCode")]
+            code: u16,
+        },
+    }
+
+    // Test struct variant with renamed fields
+    let success = Message::Success {
+        msg: "Operation completed".to_string(),
+        status_code: 200,
+    };
+
+    let json = facet_json::to_string(&success);
+    assert_eq!(
+        json,
+        r#"{"success":{"message":"Operation completed","code":200}}"#
+    );
+
+    let roundtrip: Message = facet_json::from_str(&json).unwrap();
+    assert_eq!(success, roundtrip);
+
+    // Test error variant
+    let error = Message::Error {
+        msg: "Not found".to_string(),
+        code: 404,
+    };
+
+    let json = facet_json::to_string(&error);
+    assert_eq!(
+        json,
+        r#"{"error":{"errorMessage":"Not found","errorCode":404}}"#
+    );
+
+    let roundtrip: Message = facet_json::from_str(&json).unwrap();
+    assert_eq!(error, roundtrip);
+}
+
+/// Serialization and deserialization of renamed fields in nested data structures
+#[cfg(feature = "std")]
+#[test]
+fn test_field_rename_nested_structures() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct Address {
+        #[facet(rename = "streetName")]
+        street: String,
+
+        #[facet(rename = "zipCode")]
+        postal_code: String,
+    }
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct Person {
+        #[facet(rename = "fullName")]
+        name: String,
+
+        #[facet(rename = "homeAddress")]
+        address: Address,
+
+        #[facet(rename = "contactInfo")]
+        contacts: Vec<Contact>,
+    }
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct Contact {
+        #[facet(rename = "type")]
+        contact_type: String,
+
+        #[facet(rename = "value")]
+        contact_value: String,
+    }
+
+    let person = Person {
+        name: "John Doe".to_string(),
+        address: Address {
+            street: "Main St".to_string(),
+            postal_code: "12345".to_string(),
+        },
+        contacts: vec![
+            Contact {
+                contact_type: "email".to_string(),
+                contact_value: "john@example.com".to_string(),
+            },
+            Contact {
+                contact_type: "phone".to_string(),
+                contact_value: "555-1234".to_string(),
+            },
+        ],
+    };
+
+    let json = facet_json::to_string(&person);
+    let expected = r#"{"fullName":"John Doe","homeAddress":{"streetName":"Main St","zipCode":"12345"},"contactInfo":[{"type":"email","value":"john@example.com"},{"type":"phone","value":"555-1234"}]}"#;
+    assert_eq!(json, expected);
+
+    let roundtrip: Person = facet_json::from_str(&json).unwrap();
+    assert_eq!(person, roundtrip);
+}
+
+/// Serialization and deserialization of renamed optional fields (Some and None cases)
+#[cfg(feature = "std")]
+#[test]
+fn test_field_rename_optional_values() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct OptionalFields {
+        #[facet(rename = "requiredField")]
+        required: String,
+
+        #[facet(rename = "optionalString")]
+        maybe_string: Option<String>,
+
+        #[facet(rename = "optionalNumber")]
+        maybe_number: Option<i32>,
+    }
+
+    // Test with all fields present
+    let full = OptionalFields {
+        required: "always here".to_string(),
+        maybe_string: Some("optional value".to_string()),
+        maybe_number: Some(42),
+    };
+
+    let json = facet_json::to_string(&full);
+    assert_eq!(
+        json,
+        r#"{"requiredField":"always here","optionalString":"optional value","optionalNumber":42}"#
+    );
+
+    let roundtrip: OptionalFields = facet_json::from_str(&json).unwrap();
+    assert_eq!(full, roundtrip);
+
+    // Test with None fields
+    let partial = OptionalFields {
+        required: "always here".to_string(),
+        maybe_string: None,
+        maybe_number: None,
+    };
+
+    let json = facet_json::to_string(&partial);
+    assert_eq!(
+        json,
+        r#"{"requiredField":"always here","optionalString":null,"optionalNumber":null}"#
+    );
+
+    let roundtrip: OptionalFields = facet_json::from_str(&json).unwrap();
+    assert_eq!(partial, roundtrip);
+}
+
+/// Deserialization with extra fields in JSON that aren't in the target struct
+#[test]
+fn test_field_rename_ignore_extra_fields() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct User {
+        #[facet(rename = "userId")]
+        id: u64,
+
+        #[facet(rename = "userName")]
+        name: String,
+    }
+
+    // JSON with extra fields that aren't in our struct
+    let json = r#"{"userId":123,"userName":"Alice","role":"admin","active":true}"#;
+
+    // We should be able to deserialize this without error, ignoring extra fields
+    let user: User = facet_json::from_str(json).unwrap();
+
+    assert_eq!(user.id, 123);
+    assert_eq!(user.name, "Alice");
+}
+
+/// Renamed fields have priority over original field names during serialization
+#[cfg(feature = "std")]
+#[test]
+fn test_field_rename_serialization_priority() {
+    facet_testhelpers::setup();
+
+    // When serializing, the rename attribute should always be used instead of
+    // the original field name
+    #[derive(Debug, PartialEq, Facet)]
+    struct DataModel {
+        #[facet(rename = "data")]
+        items: Vec<String>,
+    }
+
+    let model = DataModel {
+        items: vec!["one".to_string(), "two".to_string()],
+    };
+
+    let json = facet_json::to_string(&model);
+    assert_eq!(json, r#"{"data":["one","two"]}"#);
+}
+
+/// Proper errors are returned when required renamed fields are missing
+#[test]
+fn test_field_rename_missing_required_error() {
+    facet_testhelpers::setup();
+
+    #[derive(Debug, PartialEq, Facet)]
+    struct Required {
+        #[facet(rename = "renamedField")]
+        original_field: String,
+    }
+
+    // JSON missing the required field
+    let json = r#"{}"#;
+
+    // This should result in an error as the required field is missing
+    let result = facet_json::from_str::<Required>(json);
+    let e = result.unwrap_err();
+    assert!(matches!(
+        e.kind,
+        facet_json::JsonErrorKind::MissingField(f) if f == "original_field"
+    ));
+    assert_snapshot!(e.to_string());
+}
